% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geog_1kg}
\alias{geog_1kg}
\title{convenient table of thousand genome sample characteristics}
\format{data.frame}
\usage{
geog_1kg
}
\description{
convenient table of thousand genome sample characteristics
}
\keyword{datasets}
