% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_loc.R
\name{query_loc}
\alias{query_loc}
\title{query calls at a position in a Terra-based genetics cohort}
\usage{
query_loc(con, tablename = "1000_genomes_phase_3_variants_20150220", chr,
  pos)
}
\arguments{
\item{con}{BigQueryConnection instance}

\item{tablename}{character(1) table with variants}

\item{chr}{character(1) chromosome id, NCBI-like}

\item{pos}{numeric(1) position on chromosome, exact}
}
\description{
query calls at a position in a Terra-based genetics cohort
}
\note{
Metadata should be supplied.
}
